Fix argument delegation following Ruby 3 keyword arguments changes
In Ruby 3, keyword and positional arguments have been separated [docs],
which means we need to be stricter about how we're passing arguments
around.

This is possible using the ruby2_keywords gem (for Ruby 2.6), or the
`...` operator in Ruby 2.7+, but since we need to support Ruby 2.6 and
Ruby 3+, the simplest solution here is just to be stricter about what
arguments we're passing around and replace `*args` with stricter
`**kwargs` or named arguments (this is easy for us since we control this
full API anyway).

Unfortunately the tests must be slightly modified to also pass hashes
explicitly as positional arguments, for the benefit of Ruby 3 (however,
in Ruby 2.6 and 2.7 the previous code will continue to work).

[docs]: https://www.ruby-lang.org/en/news/2019/12/12/separation-of-positional-and-keyword-arguments-in-ruby-3-0/